Refactor #5 : ChatGPT AI 대화형 큐레이션 기능 구현
