+++ server/src/main/java/com/noterror/app/api/ChatGPTAIController.java
@@ -1,14 +1,29 @@
-package com.noterror.app;
+package com.noterror.app.api;
 
+import com.noterror.app.api.global.response.AiAnswerResponse;
 import org.json.JSONObject;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.util.Map;
 
-public class ChatGptApiConsole {
-    public static void chatGPT(String text) throws Exception {
+@RestController
+@RequestMapping("/api")
+public class ChatGPTAIController {
+
+    @PostMapping("/helper")
+    public ResponseEntity postChatGPTApi(@RequestBody Map<String, String> userInput) throws Exception {
+
+        String text = userInput.get("text");
+        System.out.println("text = " + text);
         String url = "https://api.openai.com/v1/completions";
         HttpURLConnection con = (HttpURLConnection) new URL(url).openConnection();
 
@@ -20,7 +35,7 @@ public class ChatGptApiConsole {
         data.put("model", "text-davinci-003");
         data.put("prompt", text);
         data.put("max_tokens", 2048);
-        data.put("temperature", 0.3);
+        data.put("temperature", 0.5);
         data.put("top_p",1);
         data.put("frequency_penalty", 0);
         data.put("presence_penalty",0);
@@ -30,20 +45,11 @@ public class ChatGptApiConsole {
 
         String output = new BufferedReader(new InputStreamReader(con.getInputStream())).lines()
                 .reduce((a, b) -> a + b).get();
+        String answer = new JSONObject(output).getJSONArray("choices").getJSONObject(0).getString("text");
 
-        System.out.println("AI : " + new JSONObject(output).getJSONArray("choices").getJSONObject(0).getString("text"));
+        System.out.println("answer = " + answer);
+        return new ResponseEntity(new AiAnswerResponse(answer), HttpStatus.OK);
     }
 
-    public static void main(String[] args) throws Exception {
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        System.out.print("나 : ");
-        String line;
-        while( (line = br.readLine()) != null) {
 
-            chatGPT(line);
-            System.out.println();
-            System.out.print("나 : ");
-        }
-        br.close();
-    }
 }
