+++ server/src/main/java/com/noterror/app/api/global/response/AiAnswerResponse.java
@@ -1,0 +1,13 @@
+package com.noterror.app.api.global.response;
+
+import lombok.Getter;
+
+@Getter
+public class AiAnswerResponse {
+
+    String answer;
+
+    public AiAnswerResponse(String answer) {
+        this.answer = answer;
+    }
+}
